+++ NEWS.md
@@ -1,5 +1,12 @@
 ## PairwiseListMatrices.jl Release Notes
 
+## Changes from v0.6 to v0.7
+
+PairwiseListMatrices v0.7 requires Julia 0.7/1.0.
+PairwiseListMatrices v0.6 was the last version with Julia 0.6 support.
+
+* `writecsv` changed to `writedlm` using `delim=','`.   
+
 ### Changes from v0.5 to v0.6
 
 PairwiseListMatrices v0.6 requires Julia v0.6.